requirements.txt
# Core ML and Computer Vision
torch>=1.13.0
torchvision>=0.14.0
ultralytics>=8.0.0
opencv-python>=4.7.0
numpy>=1.21.0
Pillow>=9.0.0

# Web Framework and API
flask>=2.3.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
gunicorn>=21.2.0
waitress>=2.1.2

# Web Interface
gradio>=4.0.0
opencv-python>=4.7.0

# Data Processing and Visualization
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.15.0

# Configuration and Environment
python-dotenv>=1.0.0
PyYAML>=6.0

# System Monitoring and Logging
psutil>=5.9.0
logging>=0.4.9.6
structlog>=23.1.0

# Security and Validation
cryptography>=41.0.0
jsonschema>=4.18.0
werkzeug>=2.3.0

# Production Monitoring
prometheus-client>=0.17.0
sentry-sdk[flask]>=1.30.0

# Testing and Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Database (optional)
sqlalchemy>=2.0.0
alembic>=1.11.0
